Refactor prediction
