+++ HeMa/src/prediction/GetterPredictor.java
@@ -1,4 +1,4 @@
-package predictor;
+package prediction;
 
 import com.github.javaparser.ast.body.FieldDeclaration;
 import com.github.javaparser.ast.body.MethodDeclaration;
@@ -11,8 +11,6 @@ import util.Tokenizer;
 import java.util.List;
 
 public class GetterPredictor extends GetterSetterPredictor {
-    public static int predicted = 0;
-    public static int correct = 0;
 
     @Override
     public boolean predict(MethodDeclaration method) {
