+++ HeMa/src/prediction/Predictor.java
@@ -1,12 +1,15 @@
-package predictor;
+package prediction;
 
 import com.github.javaparser.ast.body.MethodDeclaration;
 
-public interface IPredictor {
+public abstract class Predictor {
+    int predicted = 0;
+    int correct = 0;
+
     /**
      * Predicts the name of the method
      * @param method declaration
      * @return boolean representation of whether a prediction was successfully made
      */
-    boolean predict(MethodDeclaration method);
+    abstract boolean predict(MethodDeclaration method);
 }
